Finished first round of restructuring.
