+++ pyment/docstrings/generators/__init__.py
@@ -1,0 +1,5 @@
+"""Module for generating docstrings."""
+
+# Still way too intertwined
+
+from .numpydoc_generator import DocString
