+++ pyment/docstrings/helpers.py
