--- pyment/docstringssssss/generators/__init__.py
@@ -1,1 +1,0 @@
-"""Module for generating docstrings."""
